compiled .. at last))

+++ src/main/java/Lesson1/Athlete1.java
@@ -1,9 +1,8 @@
 package Lesson1;
 
 public class Athlete1 extends Players {
-    public Athlete1(String name, int runSpeed, int swimSpeed, int jumpHeight) {
+
+    public Athlete1(String name, double runSpeed, double swimSpeed, double jumpHeight) {
         super(name, runSpeed, swimSpeed, jumpHeight);
     }
-
-
 }
