+++ src/main/java/Lesson1/Athlete2.java
@@ -1,7 +1,7 @@
 package Lesson1;
 
 public class Athlete2  extends Players {
-    public Athlete2(String name, int runSpeed, int swimSpeed, int jumpHeight) {
+    public Athlete2(String name, double runSpeed, double swimSpeed, double jumpHeight) {
         super(name, runSpeed, swimSpeed, jumpHeight);
     }
 }
