+++ src/main/java/Lesson1/Athlete3.java
@@ -1,7 +1,7 @@
 package Lesson1;
 
 public class Athlete3 extends    Players{
-    public Athlete3(String name, int runSpeed, int swimSpeed, int jumpHeight) {
+    public Athlete3(String name, double runSpeed, double swimSpeed, double jumpHeight) {
         super(name, runSpeed, swimSpeed, jumpHeight);
     }
 }
