+++ src/main/java/Lesson1/Athlete4.java
@@ -1,9 +1,7 @@
 package Lesson1;
 
 public class Athlete4 extends  Players{
-    public Athlete4(String name, int runSpeed, int swimSpeed, int jumpHeight) {
+    public Athlete4(String name, double runSpeed, double swimSpeed, double jumpHeight) {
         super(name, runSpeed, swimSpeed, jumpHeight);
     }
-
-
 }
